+++ .gitignore
@@ -1,0 +1,4 @@
+/claro-servers/Gestion-Servers/Gestion-Server-Discovery/target/
+/claro-servers/Gestion-Servers/Gestion-Server-Zuul/target/
+/claro-servers/Gestion-Servers/Gestion-Server-Turbine/target/
+/claro-servers/Gestion-Servers/Gestion-Server-HystrixDashboard/target/
