+++ claro-servers/Gestion-Servers/Gestion-Server-Config/src/main/java/com/Gest/server/config/ConfigServerApplication.java
@@ -1,4 +1,4 @@
-package hello;
+package com.Gest.server.config;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,9 +6,9 @@ import org.springframework.cloud.config.server.EnableConfigServer;
 
 @EnableConfigServer
 @SpringBootApplication
-public class ConfigServiceApplication {
+public class ConfigServerApplication {
     
     public static void main(String[] args) {
-        SpringApplication.run(ConfigServiceApplication.class, args);
+        SpringApplication.run(ConfigServerApplication.class, args);
     }
 }
