+++ claro-servers/Gestion-Servers/Gestion-Server-Discovery/src/main/java/com/Gest/server/discovery/DiscoveryApplication.java
@@ -4,10 +4,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+//import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
 @SpringBootApplication
 @EnableEurekaServer
-@EnableDiscoveryClient
+//@EnableDiscoveryClient
 public class DiscoveryApplication {
 
 	public static void main(String[] args) {
