--- claro-servers/Gestion-Servers/Gestion-Server-Zuul/src/main/java/com/Gest/server/zuul/filters/PreFilter.java
@@ -1,46 +1,0 @@
-package com.Gest.server.zuul.filters;
-
-import com.netflix.zuul.ZuulFilter;
-import com.netflix.zuul.context.RequestContext;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import com.sun.jersey.api.client.ClientHandlerException;
-import javax.servlet.http.HttpServletRequest;
-
-/**
- *
- *
- * @author
- */
-public class PreFilter extends ZuulFilter {
-
-    private static Logger log = LoggerFactory.getLogger(PreFilter.class);
-
-    @Override
-    public String filterType() {
-        return "pre";
-    }
-
-    @Override
-    public int filterOrder() {
-        return 1;
-    }
-
-    @Override
-    public boolean shouldFilter() {
-        return true;
-    }
-
-    @Override
-    public Object run() {
-        RequestContext ctx = RequestContext.getCurrentContext();
-        HttpServletRequest request;
-        request = ctx.getRequest();
-        log.info(String.format("%s request to %s", request.getMethod(), request.getRequestURL().toString()));
-        return null;
-    }
-
-    private String getTokenHystrixFallbackMethod() {
-        return "El servicio se encuentra no disponible. Inténtelo más tarde.";
-    }
-}
